completely deleted shoppingList model because it seems unnecessary. I don't plan on allowing the user to create multiple shopping lists and it would be fairly trivial to add it back if needed
bind the shoppingListIngredients to the userId
change the deletion of recipes, ingredients and shoppingListIngredients to be soft-delete
merge recipes GET requests into one endpoint with query params
merge ingredients GET requests into one endpoint with query params

+++ apps/api/src/recipes/recipes.controller.ts
@@ -23,22 +23,10 @@ export class RecipesController {
     });
   }
 
-  @TsRestHandler(contract.recipes.getForUser)
-  async getForUser(@GetCurrentUserId() userId: number) {
-    return tsRestHandler(contract.recipes.getForUser, async () => {
-      const recipes = await this.recipesService.getForUser(userId);
-
-      return {
-        status: 200,
-        body: recipes,
-      };
-    });
-  }
-
-  @TsRestHandler(contract.recipes.getGlobal)
-  async getGlobal() {
-    return tsRestHandler(contract.recipes.getGlobal, async () => {
-      const recipes = await this.recipesService.getGlobal();
+  @TsRestHandler(contract.recipes.get)
+  async get(@GetCurrentUserId() userId: number) {
+    return tsRestHandler(contract.recipes.get, async ({ query }) => {
+      const recipes = await this.recipesService.get({ query, userId });
 
       return {
         status: 200,
@@ -67,11 +55,11 @@ export class RecipesController {
   }
 
   @TsRestHandler(contract.recipes.delete)
-  async delete() {
+  async delete(@GetCurrentUserId() userId: number) {
     return tsRestHandler(
       contract.recipes.delete,
       async ({ params: { id } }) => {
-        const recipes = await this.recipesService.delete(id);
+        const recipes = await this.recipesService.delete({ id, userId });
 
         return {
           status: 200,
